f1_race_engineer/requirements.txt
pydantic
langgraph
langgraph-cli[inmem]
langchain-core
langchain_openai
a2a-sdk
uvicorn